requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==8.0
djangorestframework
python-decouple==3.4
djangorestframework_simplejwt
drf-yasg==1.20
django-crispy-forms
django-shopping-cart